feat: clean up output from objToYaml function
This commit creates a new `cleanupYaml` function which cleans up the
output from the `objToYaml` function. Currently, the output contains
`\n` characters after every bullet point, which makes the yaml file
jarring to read. The new cleanup function detects for such cases and
removes the necessary `\n` characters.

This function was tested by using JSON.stringify on the output and
comparing that with the JSON stringified version of the output from
using utils.yamlParser.