fix: disappearing content for markdown files containing `---` in content
Currently, when we parse the front matter, we split the entire
markdown file content by `---` and we assign the element in index position
1 to be the page config, and the element in index position 2 to be
the page content.

The problem occurs when the page content also contains `---`. This commit
fixes the problem in the `frontMatterParser` function by joining all
elements from position 2 and beyond with `---`.