fix: expand definition of resource title for linking in navbar
Previously, a navbar object would only link to the resource room
if the navbar object title exactly matches the resource room title.
This commit expands the definition to allow any navbar object which
has a title that is a substring of the resource room title to be
interpreted as a resource room. This is due to the limited real
estate on the navbar, so long resource room titles are likely to be
truncated by the user.