fix: set default image alt text if null
This commit sets the `alt` attribute from the `infopic` section
to a default of `alt text` if the `alt` attribute is null - this
is because our build process displays an error message if the image
path is defined but the alt text is defined as null.

The fix is to set the `alt` attribute if it does not exist - this
way, the image renders correctly.