fix: do not attempt to intelligently infer resource room title
Previously, we tried to intelligently infer whether a nav bar object
is a resource room by checking whether it is a substring of the resource
room name. This resulted in TypeError problems when the resource room is
undefined (we cannot use the `includes` method if resource room title
is not defined).

This commit removes that check, so we need to correct the resource
room value (setting `resource_room: true`) manually, instead of
inferring it using the script.