feat: extract notification data from index.md
Previously, the homepage notification content was not being copied into the
front matter of the `index.md` file correctly. The site notification content
was previously in the `index.md` as the markdown content for v1, but for v2
we need that data in the front matter.

This commit modifies the `homepageModifier` function to add the notification
content in the front matter.