fix: swap resource room title and subtitle
An error in the V2 template resulted in the resource room title and
title fields being swapped. This means that if we were to migrate
naively (keeping resource room title as the title, and subtitle as
subtitle), then their positions in the homepage would be swapped after
migration.

The fix is to relabel the resources subtitle as title, and the resources
title as subtitle. While this is confusing semantically, what is more
important is preserving the visual structure of the existing website
on migration.