fix: do not add markdown text content if index
Currently, we use a general flow which updates a markdown file's
front matter and concatenates it with its original text content. This
works for ordinary markdown files, but not the `index.md` file since
the `index.md` file should not contain any text content. This commit
fixes that by adding an `isIndex` flag which ensures that text content
is not concatenated for the `index.md` file.