fix: use existing date from resource file name
Previously, we attempted to extract the date from the resource page's
front matter when generating the file name for the resource file
migration. This is a problem in situations where the `date` attribute
is not defined in the front matter - and in fact, Jekyll doesn't require
the `date` attribute to be defined.

To fix this, we simply re-use the existing date in the file name
to generate the new file name.