fix: swap values of infobar/infopic title and subtitle
Due to the change in template from V1 to V2, the positions of the
Title and Subtitle attributes have been swapped. This leads to a change
in the appearance of the user's homepage.

To retain the homepage's appearance when migrating, this commit
swaps the values of the title and subtitle attributes.