fix: deploy.sh bug on Netlify
Previously, the deployment script compiled a command to find all the collection.yml files in a site repository, and concatenate them into a command via a loop. However, since the default internal field separator IFS was spaces, collection.yml file paths that contained spaces, e.g. /_hello new folder/collection.yml could not be parsed as one filepath, but was parsed separately. A hacky fix was introduced to temporarily change the IFS to new lines, complete the loop, then set the IFS back to the default. This fix worked locally, but introduced a bug into the Netlify deployment, as the GNU shell does not seem to be compatible with this IFS change trick.

This PR replaces the loop with a replacement command to change from space-separated collection.yml files to comma-separated files, which Jekyll accepts as the argument to --config.

An example of Netlify deployment bug can be found here https://app.netlify.com/sites/friendly-pare-41cc65/deploys/6036fcd1c0fbf600070b9021, line 63.